+++ src/cli/src/release/mod.rs
@@ -1,8 +1,6 @@
-mod stage;
 mod publish;
 mod util;
 
-pub use stage::stage;
 pub use publish::publish;
 pub use util::{
     create_release, create_tag, extract_notes, get_remote_repo, parse_github_repo,